+++ src/graph/renderer/GridRenderer.cpp
@@ -39,7 +39,7 @@ void GridRenderer::draw_win(const RenderParams& params, MultiViewWindow* win, Re
 			}
 
 		session->drawing_helper->set_color(_color());
-		session->drawing_helper->set_line_width(line_width());
+		session->drawing_helper->set_line_width((float)line_width());
 		session->drawing_helper->draw_lines(points, false);
 	}
 }
