+++ src/graph/renderer/PointListRenderer.h
@@ -23,7 +23,7 @@ public:
 
 	void draw_win(const RenderParams& params, MultiViewWindow* win, RenderViewData& rvd);
 
-	dataflow::Setting<float> radius{this, "radius", 0.2f, "range=0:99:0.1"};
+	dataflow::Setting<double> radius{this, "radius", 0.2, "range=0:99:0.1"};
 	dataflow::Setting<color> _color{this, "color", White};
 
 	dataflow::InPort<Array<vec3>> in_points{this, "points"};
