+++ src/graph/renderer/VectorFieldRenderer.cpp
@@ -43,7 +43,7 @@ void VectorFieldRenderer::draw_win(const RenderParams& params, MultiViewWindow* 
 	}
 
 	session->drawing_helper->set_color(_color());
-	session->drawing_helper->set_line_width(line_width());
+	session->drawing_helper->set_line_width((float)line_width());
 	session->drawing_helper->draw_lines(points, false);
 }
